modify pom
