+++ sharding-orchestration/sharding-orchestration-core/pom.xml
@@ -9,4 +9,12 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>sharding-orchestration-core</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>io.shardingsphere</groupId>
+            <artifactId>sharding-orchestration-reg</artifactId>
+            <version>${sharding-sphere.version}</version>
+        </dependency>
+    </dependencies>
 </project>
