+++ sharding-orchestration/sharding-orchestration-jdbc/pom.xml
@@ -8,8 +8,13 @@
         <version>3.0.0.M3-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>sharding-orchestration-jdbc</artifactId>
 
-
+    <dependencies>
+        <dependency>
+            <groupId>io.shardingsphere</groupId>
+            <artifactId>sharding-orchestration-core</artifactId>
+            <version>${sharding-sphere.version}</version>
+        </dependency>
+    </dependencies>
 </project>
