+++ sharding-orchestration/sharding-orchestration-reg/pom.xml
@@ -8,8 +8,5 @@
         <version>3.0.0.M3-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>sharding-orchestration-reg</artifactId>
-
-
 </project>
