+++ sharding-orchestration/sharding-orchestration-reg/src/main/java/io/shardingsphere/orchestration/reg/exception/RegExceptionHandler.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.jdbc.orchestration.reg.exception;
+package io.shardingsphere.orchestration.reg.exception;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
